avoid brokenpipe, add comments, separate into headers for easy reading of code, more comments to be added later
